feat: add changing language, case and keyboard
