+++ package.json
@@ -30,5 +30,9 @@
     "webpack": "^4.41.2",
     "webpack-cli": "^3.3.9",
     "webpack-dev-server": "^3.9.0"
+  },
+  "dependencies": {
+    "auto-bind": "^3.0.0",
+    "autobind-decorator": "^2.4.0"
   }
 }
