+++ src/index.js
@@ -1,4 +1,3 @@
 import './scss/style.scss';
 import './createKeyboard.js';
 import './mouseEvent.js';
-import './keyEvent.js';
